refactory : header 기능 모듈화
