+++ frontend/src/pages/Header.js
@@ -4,45 +4,32 @@ import {
   Image,
   StyleSheet,
   Pressable,
-  Linking,
-  Dimensions,
   Text,
+  Dimensions,
 } from 'react-native'
+import { openURL } from '../utils/urlUtils'
+import { calculateDDay } from '../utils/dateUtils'
 
 const { width } = Dimensions.get('window')
 const scale = (size) => (width / 375) * size
 
 const Header = () => {
-  const openKakao = () => {
-    Linking.openURL('https://open.kakao.com/o/sAhUQNng')
-  }
-
-  const openLogo = () => {
-    Linking.openURL('https://www.kangwon.ac.kr/www/contents.do?key=2414')
-  }
-
-  // D-Day Calculation
-  const targetDate = new Date(2024, 5, 24) // Note: Month is 0-indexed in JavaScript
-  const currentDate = new Date()
-  const diff = targetDate - currentDate
-  const day = Math.ceil(diff / (1000 * 60 * 60 * 24))
-
-  // D-Day Text Logic
-  let dayText
-  if (day > 0) {
-    dayText = `종강 D-${day}`
-  } else if (day === 0) {
-    dayText = '종강'
-  } else {
-    dayText = '운행종료'
-  }
+  const dayText = calculateDDay(new Date(2024, 5, 24)) // Note: Month is 0-indexed in JavaScript
 
   return (
     <View style={styles.container}>
-      <Pressable onPress={openKakao} style={styles.button}>
+      <Pressable
+        onPress={() => openURL('https://open.kakao.com/o/sAhUQNng')}
+        style={styles.button}
+      >
         <Text style={styles.kakaoText}>카카오톡 문의</Text>
       </Pressable>
-      <Pressable onPress={openLogo} style={styles.button}>
+      <Pressable
+        onPress={() =>
+          openURL('https://www.kangwon.ac.kr/www/contents.do?key=2414')
+        }
+        style={styles.button}
+      >
         <Image
           source={require('../../assets/KNUB.png')}
           resizeMode="contain"
