+++ frontend/src/utils/urlUtils.js
@@ -1,0 +1,5 @@
+import { Linking } from 'react-native'
+
+export const openURL = (url) => {
+  Linking.openURL(url).catch((err) => console.error('Failed to open URL', err))
+}
